+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/instrument/web/TraceFilter.java
@@ -16,22 +16,17 @@
 package org.springframework.cloud.sleuth.instrument.web;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Enumeration;
-import java.util.Locale;
 import java.util.regex.Pattern;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
-import javax.servlet.ServletOutputStream;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
-import javax.servlet.WriteListener;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpServletResponseWrapper;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -139,7 +134,7 @@ public class TraceFilter extends GenericFilterBean {
 		}
 		// in case of a response with exception status a exception controller will close the span
 		if (!httpStatusSuccessful(response) && isSpanContinued(request)) {
-			Span parentSpan = spanFromRequest != null ? spanFromRequest.getSavedSpan() : null;
+			Span parentSpan = parentSpan(spanFromRequest);
 			processErrorRequest(filterChain, request, new TraceHttpServletResponse(response, parentSpan), spanFromRequest);
 			return;
 		}
@@ -163,6 +158,16 @@ public class TraceFilter extends GenericFilterBean {
 			spanFromRequest = createSpanIfRequestNotHandled(request, spanFromRequest, name, skip);
 			detachOrCloseSpans(request, response, spanFromRequest, exception);
 		}
+	}
+
+	private Span parentSpan(Span span) {
+		if (span == null) {
+			return null;
+		}
+		if (span.hasSavedSpan()) {
+			return span.getSavedSpan();
+		}
+		return span;
 	}
 
 	private void processErrorRequest(FilterChain filterChain, HttpServletRequest request,
@@ -249,7 +254,12 @@ public class TraceFilter extends GenericFilterBean {
 				log.debug("Trying to send the parent span " + parent + " to Zipkin");
 			}
 			parent.stop();
+			// should be already done by HttpServletResponse wrappers
+			SsLogSetter.annotateWithServerSendIfLogIsNotAlreadyPresent(parent);
 			this.spanReporter.report(parent);
+		} else {
+			// should be already done by HttpServletResponse wrappers
+			SsLogSetter.annotateWithServerSendIfLogIsNotAlreadyPresent(parent);
 		}
 	}
 
@@ -374,240 +384,3 @@ public class TraceFilter extends GenericFilterBean {
 	}
 }
 
-/**
- * We want to set SS as fast as possible after the response was sent back. The response
- * can be sent back by calling either of these methods.
- */
-class TraceHttpServletResponse extends HttpServletResponseWrapper {
-
-	private static final Log log = LogFactory.getLog(MethodHandles.lookup().lookupClass());
-
-	private final Span span;
-
-	TraceHttpServletResponse(HttpServletResponse response, Span span) {
-		super(response);
-		this.span = span;
-	}
-
-	@Override public void flushBuffer() throws IOException {
-		if (log.isTraceEnabled()) {
-			log.trace("Will annotate SS once the response is flushed");
-		}
-		try {
-			super.flushBuffer();
-		} finally {
-			SsLogSetter.annotateWithServerSendIfLogIsNotAlreadyPresent(this.span);
-		}
-	}
-
-	@Override public ServletOutputStream getOutputStream() throws IOException {
-		return new TraceServletOutputStream(super.getOutputStream(), this.span);
-	}
-
-	@Override public PrintWriter getWriter() throws IOException {
-		return new TracePrintWriter(super.getWriter(), this.span);
-	}
-}
-
-class TraceServletOutputStream extends ServletOutputStream {
-
-	private static final Log log = LogFactory.getLog(MethodHandles.lookup().lookupClass());
-
-	private final ServletOutputStream delegate;
-	private final Span span;
-
-	TraceServletOutputStream(ServletOutputStream delegate, Span span) {
-		this.delegate = delegate;
-		this.span = span;
-	}
-
-	@Override public boolean isReady() {
-		return this.delegate.isReady();
-	}
-
-	@Override public void setWriteListener(WriteListener listener) {
-		this.delegate.setWriteListener(listener);
-	}
-
-	@Override public void write(int b) throws IOException {
-		if (log.isTraceEnabled()) {
-			log.trace("Will annotate SS once the response is flushed");
-		}
-		try {
-			this.delegate.write(b);
-		} finally {
-			SsLogSetter.annotateWithServerSendIfLogIsNotAlreadyPresent(this.span);
-		}
-	}
-}
-
-class TracePrintWriter extends PrintWriter {
-
-	private static final Log log = LogFactory.getLog(MethodHandles.lookup().lookupClass());
-
-	private final PrintWriter delegate;
-	private final Span span;
-
-	TracePrintWriter(PrintWriter delegate, Span span) {
-		super(delegate);
-		this.delegate = delegate;
-		this.span = span;
-	}
-
-	@Override public void flush() {
-		if (log.isTraceEnabled()) {
-			log.trace("Will annotate SS once the response is flushed");
-		}
-		try {
-			this.delegate.flush();
-		} finally {
-			SsLogSetter.annotateWithServerSendIfLogIsNotAlreadyPresent(this.span);
-		}
-	}
-
-	@Override public void close() {
-		this.delegate.close();
-	}
-
-	@Override public boolean checkError() {
-		return this.delegate.checkError();
-	}
-
-	@Override public void write(int c) {
-		this.delegate.write(c);
-	}
-
-	@Override public void write(char[] buf, int off, int len) {
-		this.delegate.write(buf, off, len);
-	}
-
-	@Override public void write(char[] buf) {
-		this.delegate.write(buf);
-	}
-
-	@Override public void write(String s, int off, int len) {
-		this.delegate.write(s, off, len);
-	}
-
-	@Override public void write(String s) {
-		this.delegate.write(s);
-	}
-
-	@Override public void print(boolean b) {
-		this.delegate.print(b);
-	}
-
-	@Override public void print(char c) {
-		this.delegate.print(c);
-	}
-
-	@Override public void print(int i) {
-		this.delegate.print(i);
-	}
-
-	@Override public void print(long l) {
-		this.delegate.print(l);
-	}
-
-	@Override public void print(float f) {
-		this.delegate.print(f);
-	}
-
-	@Override public void print(double d) {
-		this.delegate.print(d);
-	}
-
-	@Override public void print(char[] s) {
-		this.delegate.print(s);
-	}
-
-	@Override public void print(String s) {
-		this.delegate.print(s);
-	}
-
-	@Override public void print(Object obj) {
-		this.delegate.print(obj);
-	}
-
-	@Override public void println() {
-		this.delegate.println();
-	}
-
-	@Override public void println(boolean x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(char x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(int x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(long x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(float x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(double x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(char[] x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(String x) {
-		this.delegate.println(x);
-	}
-
-	@Override public void println(Object x) {
-		this.delegate.println(x);
-	}
-
-	@Override public PrintWriter printf(String format, Object... args) {
-		return this.delegate.printf(format, args);
-	}
-
-	@Override public PrintWriter printf(Locale l, String format, Object... args) {
-		return this.delegate.printf(l, format, args);
-	}
-
-	@Override public PrintWriter format(String format, Object... args) {
-		return this.delegate.format(format, args);
-	}
-
-	@Override public PrintWriter format(Locale l, String format, Object... args) {
-		return this.delegate.format(l, format, args);
-	}
-
-	@Override public PrintWriter append(CharSequence csq) {
-		return this.delegate.append(csq);
-	}
-
-	@Override public PrintWriter append(CharSequence csq, int start, int end) {
-		return this.delegate.append(csq, start, end);
-	}
-
-	@Override public PrintWriter append(char c) {
-		return this.delegate.append(c);
-	}
-}
-
-class SsLogSetter {
-	static void annotateWithServerSendIfLogIsNotAlreadyPresent(Span span) {
-		if (span == null) {
-			return;
-		}
-		for (org.springframework.cloud.sleuth.Log log1 : span.logs()) {
-			if (Span.SERVER_SEND.equals(log1.getEvent())) {
-				return;
-			}
-		}
-		span.logEvent(Span.SERVER_SEND);
-	}
-}
